pyproject.toml
[tool.poetry]
name = "shop-project-test"
version = "0.1.0"
description = ""
authors = ["SabirovaOlya <sabirovaolya44@gmail.com>"]
readme = "README.md"

packages = [
    { include = "root" }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"